renaming of namespace for user q
